More notes on why libyaml installs don't work on CentOS with latest Ruby
This is due to different versions of CentOS (or other operating systems) having default versions of libyaml and libyaml-devel. For this version of Ruby, libyaml 0.1.4 is required, and yum only installs up through 0.1.2 on CentOS 5, and 0.1.3 on CentOS 6. This is fixed by allowing RVM to manage libyaml and specifying the libyaml when installing Ruby.